# public/v1-api.openapi.yaml
openapi: 3.0.3
info:
  title: voidanchae v1 Partner API
  version: "1.0.0"
  description: |
    파트너사 (e.g. 스테이폴리오) 가 voidanchae hosting-manager 와 양방향
    통합할 때 사용하는 외부 API. server-to-server 호출 전용.

    ## 인증

    모든 요청은 `Authorization: Bearer <API_KEY>` 헤더 필요.
    키는 `vd_live_` 또는 `vd_test_` prefix 로 시작. 발급은 voidanchae 측 관리자가 수행.

    ## 스코프

    각 키는 개별 scope 목록을 가진다. scope 가 부족하면 `403 Forbidden`:
      - `properties:read` — 지점 정보 + 가용성 조회
      - `bookings:read`   — 예약 조회
      - `cleanings:read`  — 청소 조회
      - `cleanings:write` — 청소 생성/수정/삭제

    파트너 키는 특정 지점들에만 제한될 수 있다 (ApiClient.propertyIds).
    범위를 벗어난 지점 접근 시 `403`.

    ## 응답 형식

      - 목록: `{ "items": [...] }` (향후 pagination 추가 여지)
      - 단건: 객체 직접 반환
      - 에러: `{ "error": "<Category>", "code": "<machine_code>", "details"?: [...] }`

    ## 충돌 처리 (청소)

    동일 (propertyId, date) 슬롯에 두 source 가 동시에 청소를 생성하려 하면
    먼저 생성된 쪽이 유지되고, 두번째는 `409 Conflict` 받으면서 `existing` 필드로
    기존 청소 정보를 함께 반환. 같은 파트너가 같은 `externalId` 로 재전송하면
    멱등 update (200 OK).
servers:
  - url: https://voidanchae.com
    description: production

security:
  - BearerAuth: []

paths:
  /api/v1/properties:
    get:
      summary: 지점 목록
      description: 파트너 시스템과 ID 매핑할 때 사용. 키가 propertyIds 로 제한돼 있으면 그 지점들만.
      tags: [properties]
      security: [{ BearerAuth: [properties:read] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  properties:
                    type: array
                    items: { $ref: '#/components/schemas/Property' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /api/v1/properties/{propertyId}/availability:
    get:
      summary: 일자별 가용성
      tags: [properties]
      security: [{ BearerAuth: [properties:read] }]
      parameters:
        - $ref: '#/components/parameters/PropertyIdPath'
        - in: query
          name: from
          required: true
          schema: { type: string, format: date, example: "2026-06-01" }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date, example: "2026-06-10" }
          description: from 포함 to 미포함 (hotel night convention)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  propertyId: { type: string }
                  from: { type: string, format: date }
                  to: { type: string, format: date }
                  days:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        available: { type: boolean }
                        bookingId: { type: string, nullable: true, description: 점유 중일 때 우리 측 예약 ID }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: property_not_found }

  /api/v1/bookings:
    get:
      summary: 예약 목록
      description: |
        Event (OTA mirror) + Booking (직접 예약) 통합 응답. Stayfolio 측 예약은
        Beds24 sync 를 통해 `source: 'stayfolio'` 또는 channelId 매칭으로 노출.
      tags: [bookings]
      security: [{ BearerAuth: [bookings:read] }]
      parameters:
        - in: query
          name: propertyId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/BookingStatus' }
        - in: query
          name: source
          schema: { type: string, example: stayfolio }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /api/v1/bookings/{id}:
    get:
      summary: 예약 단건
      tags: [bookings]
      security: [{ BearerAuth: [bookings:read] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: booking_not_found }

  /api/v1/cleanings:
    get:
      summary: 청소 목록
      tags: [cleanings]
      security: [{ BearerAuth: [cleanings:read] }]
      parameters:
        - in: query
          name: propertyId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/CleaningStatus' }
        - in: query
          name: externalSource
          schema: { type: string, description: 본인이 생성한 청소만 필터 시 사용 (e.g. 'stayfolio') }
        - in: query
          name: limit
          schema: { type: integer, default: 200, maximum: 500 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Cleaning' }

    post:
      summary: 청소 생성
      description: |
        파트너 시스템의 청소 일정을 voidanchae 로 push. `externalId` 는 파트너
        측의 고유 ID — 같은 ID 재전송 시 idempotent update.

        같은 (propertyId, date) 슬롯이 이미 점유된 경우(우리 admin 또는 다른
        파트너가 먼저 생성) `409 Conflict`. `existing` 필드로 기존 데이터 확인.

        담당 청소부는 두 방식 중 하나:
          - `cleanerId` (voidanchae 내부 Cleaner.id) — 공유 풀 사용
          - `externalCleanerName` + `externalCleanerPhone` — 파트너 자체 풀
      tags: [cleanings]
      security: [{ BearerAuth: [cleanings:write] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CleaningCreate' }
      responses:
        '200':
          description: 기존 (externalSource, externalId) 매칭 → update 결과
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cleaning' }
        '201':
          description: 신규 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cleaning' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: slot_already_claimed — 같은 슬롯이 다른 source 로 점유됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  code: { type: string }
                  existing: { $ref: '#/components/schemas/Cleaning' }

  /api/v1/cleanings/{id}:
    get:
      summary: 청소 단건
      tags: [cleanings]
      security: [{ BearerAuth: [cleanings:read] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cleaning' }
        '404': { description: not_found }

    patch:
      summary: 청소 수정 (본인 것만)
      description: |
        본인 파트너가 만든 청소(externalSource 일치)만 수정 가능. 다른 파트너
        또는 내부 청소는 `403 not_your_cleaning`.

        `status` 가 'done' 으로 전환되면 `completedAt` 자동 세팅.
      tags: [cleanings]
      security: [{ BearerAuth: [cleanings:write] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CleaningPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cleaning' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: not_your_cleaning }
        '404': { description: not_found }

    delete:
      summary: 청소 삭제 (본인 것만)
      tags: [cleanings]
      security: [{ BearerAuth: [cleanings:write] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        '403': { description: not_your_cleaning }
        '404': { description: not_found }

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: vd_live_*

  parameters:
    PropertyIdPath:
      in: path
      name: propertyId
      required: true
      schema: { type: string }

  responses:
    BadRequest:
      description: 입력 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 인증 실패 (헤더 누락, 잘못된 키, 만료/회수)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 권한 부족 (scope, propertyId 범위, 소유권)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string, example: BadRequest }
        code:  { type: string, example: invalid_body }
        details:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }

    Property:
      type: object
      properties:
        id: { type: string }
        name: { type: string, example: 안온재 }
        timezone: { type: string, example: Asia/Seoul }
        beds24PropId: { type: string, nullable: true }
        maxGuests: { type: integer, nullable: true }

    BookingStatus:
      type: string
      enum: [confirmed, inquiry, cancelled, pending]

    Booking:
      type: object
      properties:
        id: { type: string, description: voidanchae 내부 ID }
        source: { type: string, example: stayfolio }
        externalId: { type: string, nullable: true, description: OTA/파트너 측 예약 ID }
        propertyId: { type: string }
        guestName: { type: string }
        guestEmail: { type: string, nullable: true }
        guestPhone: { type: string, nullable: true }
        numAdults: { type: integer }
        numChildren: { type: integer }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        status: { $ref: '#/components/schemas/BookingStatus' }
        createdAt: { type: string, format: date-time }

    CleaningStatus:
      type: string
      enum: [pending, done]

    Cleaning:
      type: object
      properties:
        id: { type: string }
        propertyId: { type: string }
        date: { type: string, format: date }
        status: { $ref: '#/components/schemas/CleaningStatus' }
        cleanerSource:
          type: string
          enum: [internal, external]
          nullable: true
          description: internal = voidanchae Cleaner 풀, external = 파트너 자체 풀
        cleanerName: { type: string, nullable: true }
        cleanerPhone: { type: string, nullable: true }
        externalSource: { type: string, nullable: true, example: stayfolio }
        externalId: { type: string, nullable: true }
        supplies: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        completionNote: { type: string, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        hasIssue: { type: boolean }
        createdAt: { type: string, format: date-time }

    CleaningCreate:
      type: object
      required: [propertyId, date, externalId]
      properties:
        propertyId: { type: string }
        date: { type: string, format: date }
        externalId:
          type: string
          description: 파트너 측 고유 ID — 멱등성 키
        externalCleanerName: { type: string, nullable: true, description: 파트너 자체 풀 사용 시 }
        externalCleanerPhone: { type: string, nullable: true }
        cleanerId: { type: string, nullable: true, description: voidanchae 공유 풀 사용 시 }
        status: { $ref: '#/components/schemas/CleaningStatus' }
        supplies: { type: string, nullable: true }
        notes: { type: string, nullable: true }

    CleaningPatch:
      type: object
      description: 1개 이상 필드 필수
      properties:
        externalCleanerName: { type: string, nullable: true }
        externalCleanerPhone: { type: string, nullable: true }
        cleanerId: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/CleaningStatus' }
        supplies: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        completionNote: { type: string, nullable: true }